+++ near-sdk/src/collections/vector.rs
@@ -265,60 +265,12 @@ mod tests {
     use borsh::BorshDeserialize;
     use rand::{Rng, SeedableRng};
 
-    use near_vm_logic::types::AccountId;
-    use near_vm_logic::VMContext;
-
     use crate::collections::{next_trie_id, Vector};
-    use crate::{env, MockedBlockchain};
-
-    fn alice() -> AccountId {
-        "alice.near".to_string()
-    }
-
-    fn bob() -> AccountId {
-        "bob.near".to_string()
-    }
-
-    fn carol() -> AccountId {
-        "carol.near".to_string()
-    }
-
-    fn set_env() {
-        let context = VMContext {
-            current_account_id: alice(),
-            signer_account_id: bob(),
-            signer_account_pk: vec![0, 1, 2],
-            predecessor_account_id: carol(),
-            input: vec![],
-            block_index: 0,
-            block_timestamp: 0,
-            account_balance: 0,
-            account_locked_balance: 0,
-            storage_usage: 10u64.pow(6),
-            attached_deposit: 0,
-            prepaid_gas: 10u64.pow(18),
-            random_seed: vec![0, 1, 2],
-            is_view: false,
-            output_data_receivers: vec![],
-            epoch_height: 0,
-        };
-        let storage = match env::take_blockchain_interface() {
-            Some(mut bi) => bi.as_mut_mocked_blockchain().unwrap().take_storage(),
-            None => Default::default(),
-        };
-        env::set_blockchain_interface(Box::new(MockedBlockchain::new(
-            context,
-            Default::default(),
-            Default::default(),
-            vec![],
-            storage,
-            Default::default(),
-        )));
-    }
+    use crate::test_utils::test_env;
 
     #[test]
     fn test_push_pop() {
-        set_env();
+        test_env::setup();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(0);
         let mut vec = Vector::default();
         let mut baseline = vec![];
@@ -336,7 +288,7 @@ mod tests {
 
     #[test]
     pub fn test_replace() {
-        set_env();
+        test_env::setup();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(1);
         let mut vec = Vector::default();
         let mut baseline = vec![];
@@ -361,7 +313,7 @@ mod tests {
 
     #[test]
     pub fn test_swap_remove() {
-        set_env();
+        test_env::setup();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(2);
         let mut vec = Vector::default();
         let mut baseline = vec![];
@@ -387,7 +339,7 @@ mod tests {
 
     #[test]
     pub fn test_clear() {
-        set_env();
+        test_env::setup();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(3);
         let mut vec = Vector::default();
         for _ in 0..100 {
@@ -403,7 +355,7 @@ mod tests {
 
     #[test]
     pub fn test_extend() {
-        set_env();
+        test_env::setup();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(0);
         let mut vec = Vector::default();
         let mut baseline = vec![];
@@ -428,7 +380,7 @@ mod tests {
 
     #[test]
     fn test_debug() {
-        set_env();
+        test_env::setup();
         let mut rng = rand_xorshift::XorShiftRng::seed_from_u64(4);
         let prefix = next_trie_id();
         let mut vec = Vector::new(prefix.clone());